backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching & Queue
redis==5.0.1
celery==5.3.6

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
flake8==7.0.0
black==24.1.1
mypy==1.8.0

# File Upload & Storage
python-magic==0.4.27
python-magic-bin==0.4.14  # For Windows

# Text Extraction
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0

# NLP & Skill Extraction
spacy==3.7.2

# AI Providers (Phase 4)
groq==0.4.2
aiohttp==3.9.3
nest-asyncio==1.6.0

# AI/ML (will be added in Phase 4)
# langchain==0.1.0
